Rename package to pyramid_resourceful
The old name was derived from the Restler package, which was a
contraction of RESTful and controller.

Restler was built on top of Pylons and the initial version of
pyramid_restler was more or less just a port of that idea to Pyramid.

v2 of pyramid_restler builds on those ideas but is significantly
different and, in my opinion, is much improved.

The new name uses "resourceful" instead of "RESTful" because,
"resourceful" is more generic, "RESTful" is overused, and this package
isn't strictly RESTful.

NOTE: "Resource" here refers to web resources in general and not
specifically to Pyramid's concept of resources.